Reset latch after each run
The throughput tests use a CountDownLatch to determine when a run
is complete. Each test performs multiple runs. Previously, the latch
was being setup once per test so for anything but the first run in the
test the latch would already be counted down.

Update the tests to reset the latch before each run.