#7 Added debug-level logging when no Selectors are matched against the given key.
Also removed some vestigal references to cache freshening code that didn't work after the change of `notify` to accept Object. When a `refreshRequired` is noted, the cache is now naively dumped and has to be re-populated key by key. It's impossible to iterate over the existing Registrations and re-populate the cache because the key coming in might be different than the one in the cache. We simply take a naive approach to this and dump everything, letting `find(Object)` re-populate the cache for the given entry.