#12 Add a Tuple class and hierarchy of type-safe subclasses to provide an easy way to send multiple arguments to a Consumer, Function, Composable, etc..., which only accept a single parameter.
The Tuple base class has helper methods on it to create, from a given set of values, a Tuple of any size. There are 8 subclasses of Tuple that provide type-safe signatures for up to the first 8 slots. After 8, the remaining values are lumped into a single Tuple of undetermined size.

This first draft version has no type coercion in it, so to extract an item of the correct type requires having set that type at the time of creation.

Tuples are backed by an immutable List (though the items placed in them might not be immutable themselves).