Remove static state. Tidy up Dispatcher lifecycle
Remove all static state from Reactor. This has resulted in Context
being deleted entirely. Dispatchers are now created directory. New
default constructors have been added to ease the creation of the
different types of Dispatcher in their default configurations.

The ReactorFactoryBean has been removed as it relied heavily upon the
state that was previously being held in R. The new direction is that
in a Spring app the application context will maintain the state.
An XML namespace and Java configuration support will be added in
the future.

Dispatcher lifecycle

The R class has been drastically reduced in size: it now only has
a single method on it.