A new type of Reactor that can act as a TCP server
An first cut of a TcpServerReactor that can act as a TCP server,
listening on a specified port. Appropriately registered consumers are
invoked when requests are received by the server. Consumers can send
replies via the replyTo key in the Event that they have
accepted.

This functionality is still very much subject to change.