Rework Environment configuration
Update the format of the properties file that's used to configure an
environment to allow multiple dispatchers of the same type to be
configured at the same time. To enable this, a dispatcher's type has
been separated from its name. This has the benefit that a new type of
dispatcher can be slotted in, without having to change any code: it's
simply a matter of updating the configuration.