Simplify Observable's generic type signatures
Observable declared a number of methods with <T, E extends Event<T>>
and then didn't reference T in the method's arguments. Therefore,
the signature can be simplified to <E extends Event<?>>. In addition
to this signature simplification, Observable's javadoc has been
updated to reflect the generic type changes and for general
consistency.