Remove Reactor registration methods
The use of Registry and the resulting exposure of Registrations from
Environment brought a number of problems with it. For example,
Registrations can be paused and can be cancelled after use which does
not make sense for a Reactor registered with an Environment.

The ability to register Reactors with an Environment was not being
used anywhere. Taking this into account, along with the design issues
described above, the decision has been made to remove this
functionality. It is expected to return in a better, more considered
form in the future when a concrete need arises.

Fixes #107