Remove concept of Taggable selectors and keys
The tag-aware selection strategy was intended to make the keys that a
selector would match dynamic. A match would only be made if the selector
and the key had one or more tags in common. In reality this
functionality was not being used anywhere and the implementation had
some limitations that we're not happy with.

Until there's a concrete use-case for the functionality, and we can
use that to drive a better design, we've decided to remove the tag gable
concept.

Fixes #112