Remove SelectionStrategy now that it's not implemented
In #112, TaggableSelectionStategy was removed. This was the only
implementation of SelectionStrategy so the interface is now
unnecessary. It has been removed and the code that still referenced
it has been updated. CachingRegistry will now always use its
default selection strategy which is to call the match method of the
registered Selectors.

Fixes #131