Don't apply Eclipse plugin to top-level project
If the Eclipse plugin is applied to the top-level project, after
running ./gradlew eclipse, Eclipse's import existing projects
functionality will only see a single project to import.

The apply of the plugin has been moved so that it's only applied to
subprojects. This allows Eclipse's import to see all of the projects.