Introduced a Sequencer for ordered and optionally persistent sequencing of events. It ensures that all events published on a key are ordered through a pluggable Queue implementation. The default is to use the same type of Queues for other places within Reactor (the LinkedTransferQueue e.g.).
A Sequencer can be configured to persist events by setting it's Queue to be an instance of one of the newly-introduced PersistentQueue implementations. There are two default impls: InMemory and IndexedChronicle. The former uses a Map and the latter uses a JavaChronicle IndexedChronicle to persist Events to the filesystem. It should be fairly straightforward, then, to create an instance of a PersistentQueue pointing to the same JavaChronicle in another process and read items from Queue that were published in the first process.