Addresses #155 by adding awareness of a new CancelConsumerException to the EventRouter so that if a Consumer throws this Exception, it is caught by the EventRouter and the Registration is immediately cancelled.
Also added a CancellableConsumer which takes a Predicate in the constructor. When this Consumer is used in place of a standard Consumer, the user can implement the doAccept(T) method and be assured that it will only be called while the Predicate is true. If the Predicate returns false, the Consumer throws a CancelConsumerException so that it can be removed from the Registry.