Fix problem where closing a TcpClient may timeout
Intermittently, closing a TcpClient will timeout. This will result in
await() on the close Promise blocking for the configured timeout
period.

The problem appears to be due to NettyTcpClient duplicating logic in
TcpClient that closes all of the client's TcpConnections and
performing the duplicated closes after it's initiated graceful shutdown
of the EventLoopGroup. This sets up a race between the event loop
shutting down and the connections being closed. If the EventLoopGroup
shutdown wins the race, the connection close will fail with a
RejectedExecutionException which appears to stall the shutdown of the
TcpClient such that the Promise is not fulfilled in a timely manner.

To address this problem, the duplicated close logic has been removed so
that the closing of the connections is only performed before the
EventLoopGroup has begun its shutdown processing. This should hopefully
address the problem with TcpClientTests.closingPromiseIsFulfilled()
taking 30 seconds for the close Promise's success consumer to be called.