Only reconnect when connection was dropped
Previously, no distinction was made between a connection that had
been dropped and a connection that was closed. This meant that when
a TcpConnection was closed it would attempt to reconnect. The same
was true of a TcpClient being closed: all of its TcpConnections would
attempt to connect during the close processing. This latter case
resulted in exceptions being thrown as Netty was in the process of
shutting down while the connections were being re-established.

NettyTcpConnection now keeps track of whether or not it's closing, i.e.
whether or not it's been explicitly closed rather than being dropped.
If it is closing then no attempt is made to reconnect. Similarly, 
NettyTcpClient also keeps track of whether or not it's closing. If it
is, it's connection Supplier will not attempt to create a new connection
and will return null instead.

A unit test has been added to verify that a dropped connection will
still attempt to reconnect.