Changed UUID generation strategy to eliminate eaio as an external dependency.
After recent internal discussion around UUID generation, it was pointed out that including the MAC address of a node into the event's ID may not be a desirable use of that information. It's also not clear what definite advantage using the MAC address gives in practical application.

Unless we can demonstrate differently, it's probably a good idea to err on the side of eliminating a dependency and replace generating the nodeAndClockSeq value with a seeded, SecureRandom long value generation, which is what this change does.