Added Jackson + Afterburner JsonPathSelector implementation for matching JsonPath expressions against key objects.
Internally this is implemented by providing an intelligent JsonProvider for Jackson 2. This means the JsonPathSelector can match against JSON data in the form of a String, a byte[], a Map, a List, or a JsonNode read from ObjectMapper.readTree(json).

Initial benchmark performance tests show about 1.2M selections/sec on a new MacBook Pro. That's considerably less than the ObjectSelector, which can get as much as 18-20M selectors/sec. But the flexibility of being able to assign Consumers based on the values extracted from a single JSON document is probably worth the trade-off if you need that functionality.