Fixes #211 Guards against double dispatch when handling errors.
Previously, a Reactor's internal errorHandler would try to dispatch the error handling into the configured Dispatcher. If every Consumer throws a RuntimeException of some kind, then the Dispatcher would be swamped with twice the number of requests--one for the event itself, and another for the exception handler.

This change selects the error handlers directly from the Consumer Registry and invokes them in the calling thread which will always be the Dispatcher's thread of execution since the errorHandler is only ever invoked by the Dispatcher code in a catch block. This eliminates the double dispatch and causes the error handler to function normally.