#241 #245 Adds a couple of constructor parameters to Reactor and a couple of methods to EventRoutingComponentSpec for specifying two types of error handlers and the Selector to be used as a default.
Error handlers are split into two different kinds: the dispatchErrorHandler is what is passed to the Dispatcher when dispatching events and will be called in case of an Exception being raised from calling a Consumer's accept method. The uncaughtErrorHandler, however, will be invoked whenever an error is reported on a Reactor using the notify(Exception.class, Event) method. This is useful for dealing with errors that occur outside this Reactor's Consumers (like in a TCP server or somesuch).

Also adds the ability to specify the Selector to use as a default in case that object needs to be reused in several places.

Sensible defaults are used if these additional parameters are null.