Initial sketch of generic object pool
This implements pooling of Events. Events are pooled based
on their generic type.

Pool is implemented through the bit-set, since it should yield a better
performance than maintaining and synchronising access to two lists.

When next freely available object is required, bit-set is queried for
the next free bit position. If the position is larger than whatever the
pool has so far, new instance is created, pooled and returned.
Objects are never released.