Added support for Kryo, Protocol Buffers, and Jackson JSON support based around a shared SerializationCodec.
The new abstract SerializationCodec handles writing type names into the Buffer so that any object can be read and written without registering any special handlers. It does this by encoding into the Buffer the length of the class name as an int, then the class name as bytes, then the actual bytes of the message. When decoding, the process is reversed and the class name is read from the Buffer so that that exact class type can be instantiated. Most serialization engines require a type to be efficient anyway, so the only thing needed in the Codec is to maintain the class name of the original object.